src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
fonttools==4.28.5
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
